sage: allow device specific image override; stop copying useless images.
we used to copy both sets of sage binaries (_dev and not) on the device
but in reality one only ever need one set for a given device, so save
the copying and image space.

additionally we allow overwrite of the base path configuration for the
binaries to be pulled into image; this allows to customize test binaries
more easily and it also allows customer to properly select the set they
need (_dev or not) depending on their actual part.

all the reference designs are using _dev target parts.

refs #SWANDROID-3703

Change-Id: I78b273ba223475e64ead094e0c64c6e252a3af99
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>