device: advertise useful middleware libraries as public.
with the now default early loading of graphics library at activity
start time that android offers, the process issues a nexus join connection
from the zygote system scope at application loading time (start of the
activity).

when the application is also providing a jni component making use of nexus
later on, the run time would cause a potential reload of middleware libraries
from within the application private scope due to the namespace restrictions
enforcement of android.  however, doing so would also duplicate the existing
nexus local ipc process state that was created|initialized at activity launch
time and therefore would cause a new nexus join to be issued to the server.
the latter would catch the unbalanced client state on 'double join' and assert,
rightfully, but annoyingly for the rest of the system. due to the duplication
of the nexus local ipc process state, there is no safeguard to ensure nexus
join is not called multiple times within the same process scope.

to prevent this situation from occuring, we now need to make the middleware
libraries of interest 'public' which would prevent the need to reload them
in the application private scope and which would therefore not create a
duplicate (and invalid) nexus local ipc process state, eliminating the server
crash.

refs #SWANDROID-4196

Change-Id: I0932b4b0f790e4b2f981af9d4d3aec4c34b4daee
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>