device: rename nexus-memory device to "nxmem".
we used to name the nexus-memory device "ion" to match existing entry in the
android exception list for devices; this in turns caused other aspects
of the system to inherit the 'true' ion-device policies (such as selinux);
and as such we ended up with potential flooding of selinux audit which may
cause hiccups during video playback when the kernel dmesg are at default
level. now in O, such exception list (and enforcement thru cts) no longer
applies.

rename the nexus-device to something more prorietary and properly allocate
policies for it.

refs #SWANDROID-4393

Change-Id: Ie1c0991e1b6de359bbe6c26af7039711ad443a1a
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>