kernel: 4.1|4.9 compatibility.
pass the keyring in the new config option (in 4.1 this is a no-op).
at this time, we only pass the verity key, we will need to pass in the
whole keyring with private signing per device as well in the final
integration.

refs #SWANDROID-4442

Change-Id: I377d5286f3f798a7b650828c4584cc9650a47530
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>
(cherry picked from commit 476a2e5756227f8c9e210401f0f2627e9b389b7c)