sepolicy: support flavor of hwcfg based on bp3 support.
for parts enabling bp3 feature set, we need to change the
hwcfg partition from cramfs|readonly to vfat|readwrite.

parts supporting bp3 are starting with 7271|7268 chipsets
and not all parts can necessarily support it.  a further
complication is that for some existing field devices
deployment, changing from cramfs to vfat requires some
manipulation of the hwcfg content which may not be obvious
to achieve without causing problems on the device.

for those reasons, we currently only enable this function on
demand and allow to maintain the former model.

refs #SWANDROID-4863

Change-Id: I6a8c726c30a07afcb5219a10e02255f0190038d4
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>
(cherry picked from commit 3647f611d47a321c83cfee7101a6d59746051e92)