manifest:treble: run allocator interface in passthrough transport.
this is a workaround.

when running the allocator interface in hwbinder, the device may lock
up under some circumstances to be determined where the binder gets
held up and the interface is never invoked.  this lead to a lock up
and anr.  it is not clear at the moment why this would happen, in the
mean time, running the interface in passthrough avoids the problem.

refs #SWANDROID-4578

Change-Id: Ibab44860f069423978782309439647d72bd1192c
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>