device: allow bindist bootloader images even if source code is present.
This changes allows OEM's and CAS vendors to build the Android-P tree with
pre-built bootloader binaries that are held in a OEM/CAS bindist dir
within the tree.

The variable "BCM_DIST_BLIMG_BINS" needs to be set to "y" to prevent the
existing BOLT code from being built (even if it is present in the tree).
There is no need to delete or move the original bolt code.

refs #SWANDROID-5360

Change-Id: I42c6f7b93ae1ccdf95742d219bd91d0d1070e4b0
Signed-off-by: Rob McConnell <robert.mcconnell@broadcom.com>