device: use keymaster type1 cert from custom oem identifier for zb target.
the custom oem identifier is 0x2b08 and is matching the zb devices deployed
in particular at google, since this fallback to type1 is implemented mostly
to allow seamless upgrade of existing devices to keymaster services, it is
selected by default for the image.

the bcm oem identifier (0x1111) reference zb type1 cert is kept around to
be used if necessary (can always be pushed to the device).

refs #SWANDROID-1158

Change-Id: I059518b71c367d6ec3d4105ae7b7171ec145b6fd
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>