common: add sepolicy and build support for sfhak
For MediaCas a new SfHak HAL service is provided to allow the A/V
PID channels in one process to be attached to keyslots in another
process.  This is part 1 of 2 changes, that allows this service to
be built only if ANDROID_SUPPORTS_MEDIACAS is defined in the
product .mk file.   The change also adds sfhak SELinux policies.

refs #SWANDROID-5365

Change-Id: Ic59235a682300be6f621ed8a252546df16945b59
Signed-off-by: Rob McConnell <robert.mcconnell@broadcom.com>