audio: Do not connect tunnel output to speaker device
Connecting the tunnel output to the SPEAKER device results in
any opened tunnel outputs to remain open even when HDMI cable
is removed (and the DIGITAL device becoming unavailable).
Subsequently on HDMI insertion, the profiles of the DIGITAL
device do not get re-enumerated due to the output already
being opened.

refs #SWANDROID-5557

Change-Id: I3a70f96053bdb64afba9c45731984bd1f374a0dc
Signed-off-by: Edwin Chan <edwin.chan@broadcom.com>