device: adjusted layouts for compliance.
compliance testing validates that the userdata partition is
no less than 4GB (full). per SWANDROID-4952, we added a 128MB
reserved region for critical process which does not show into
the /data accounting, causing validation of 4GB compliance for
"applications" to fail.

expanded the userdata to account for 4GB+reserved (reserved is
128MB plus placement overhead) and so properly meet expectations
that /data is seen as a full 4GB wide.  the "reserved" region
addition is absorbed by mainly reducing /vendor and /system.

refs #SWANDROID-5603
refs #SWANDROID-4952

Change-Id: Iec4c9af730c4220967fa304747f2d9f85886e6df
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>