device: adding "nvi" configuration fragment.
"nvi" (or non-vendor-image) device are pre-O (so pre-treble)
devices mode which need to disable a few functionality of android,
in particular the vndk support and linker namespace.

as a side effect of this configuration, we also may have to specify
which additional packages are required to be part of the main
/system/lib, packages which would normally only be present in the vndk
and which would not be available to the linker consequently.

nvi targets must include this fragment.

refs #SWANDROID-6085

Change-Id: Ieabf87d9151cbf52f2f21a4c42d4f553388ce0ea
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>
(cherry picked from commit 0c600962c6f62bd84cab6ae7b940d700ae9f9a06)